+++ talker/lib/src/models/models.dart
@@ -1,1 +1,2 @@
+export 'filter/filter.dart';
 export 'talker_data/talker_data.dart';