+++ talker_flutter/lib/src/models/flutter_talker_data_interface.dart
@@ -1,0 +1,4 @@
+import 'package:talker_flutter/talker_flutter.dart';
+
+abstract class FlutterTalkerDataInterface
+    implements HaveFlutterColorInterface, TalkerDataInterface {}