+++ talker_flutter/lib/src/models/flutter_talker_error.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+import 'package:talker_flutter/talker_flutter.dart';
+
+class FlutterTalkerError extends TalkerError
+    implements HaveFlutterColorInterface {
+  FlutterTalkerError(Error error, {this.color}) : super(error);
+
+  @override
+  final Color? color;
+}