+++ talker_flutter/lib/src/src.dart
@@ -1,3 +1,5 @@
 export 'talker_screen.dart';
 export 'extensions/extensions.dart';
 export 'widgets/widgets.dart';
+export 'models/models.dart';
+export 'options/options.dart';