+++ talker_flutter/lib/src/widgets/cards/talker_data_card.dart
@@ -6,8 +6,10 @@ class TalkerDataCard extends StatelessWidget {
     Key? key,
     required this.data,
     required this.onTap,
+    required this.options,
   }) : super(key: key);
 
+  final TalkerScreenOptions options;
   final TalkerDataInterface data;
   final Function() onTap;
 
@@ -24,9 +26,7 @@ class TalkerDataCard extends StatelessWidget {
             margin: const EdgeInsets.all(5).copyWith(bottom: 0),
             padding: const EdgeInsets.all(10),
             decoration: BoxDecoration(
-              border: Border.all(
-                color: data.logLevel.color,
-              ),
+              border: Border.all(color: _color),
             ),
             child: Row(
               crossAxisAlignment: CrossAxisAlignment.start,
@@ -40,7 +40,7 @@ class TalkerDataCard extends StatelessWidget {
                         text: TextSpan(
                           text: 'Message  ',
                           style: TextStyle(
-                            color: data.logLevel.color,
+                            color: _color,
                             fontWeight: FontWeight.bold,
                           ),
                           children: [
@@ -57,7 +57,7 @@ class TalkerDataCard extends StatelessWidget {
                         text: TextSpan(
                           text: 'Time  ',
                           style: TextStyle(
-                            color: data.logLevel.color,
+                            color: _color,
                             fontWeight: FontWeight.bold,
                           ),
                           children: [
@@ -75,7 +75,7 @@ class TalkerDataCard extends StatelessWidget {
                           text: TextSpan(
                             text: 'StackTrace  ',
                             style: TextStyle(
-                              color: data.logLevel.color,
+                              color: _color,
                               fontWeight: FontWeight.bold,
                             ),
                             children: [
@@ -93,7 +93,7 @@ class TalkerDataCard extends StatelessWidget {
                           text: TextSpan(
                             text: 'Additional  ',
                             style: TextStyle(
-                              color: data.logLevel.color,
+                              color: _color,
                               fontWeight: FontWeight.bold,
                             ),
                             children: [
@@ -116,7 +116,7 @@ class TalkerDataCard extends StatelessWidget {
                     onPressed: onTap,
                     icon: Icon(
                       Icons.copy,
-                      color: data.logLevel.color,
+                      color: _color,
                     ),
                   ),
                 )
@@ -127,12 +127,12 @@ class TalkerDataCard extends StatelessWidget {
             left: 15,
             child: Container(
               transform: Matrix4.translationValues(0, -8, 0),
-              color: Theme.of(context).scaffoldBackgroundColor,
+              color: options.backgroudColor,
               padding: const EdgeInsets.symmetric(vertical: 4, horizontal: 10),
               child: Text(
                 title,
                 style: TextStyle(
-                  color: data.logLevel.color,
+                  color: _color,
                   fontSize: 16,
                   fontWeight: FontWeight.bold,
                 ),
@@ -144,5 +144,12 @@ class TalkerDataCard extends StatelessWidget {
     );
   }
 
-  String _getTitle(TalkerDataInterface data) => data.displayTitle;
+  Color get _color {
+    if (data is HaveFlutterColorInterface) {
+      return (data as HaveFlutterColorInterface).color ?? data.logLevel.color;
+    }
+    return data.logLevel.color;
+  }
+
+  String _getTitle(TalkerDataInterface data) => data.displayTitleWithTime;
 }