+++ talker_flutter/lib/src/widgets/widgets.dart
@@ -1,1 +1,2 @@
 export 'cards/cards.dart';
+export 'filter/filter.dart';