+++ talker/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.6.0
+- Implement filter for logs
+
+## 0.5.2
+- Add title for default models
+
 ## 0.5.1
 - Fix README images urls